+++ aws_emr_launch/constructs/lambdas/__init__.py
@@ -11,11 +11,9 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-from .emr_profile import (
-    EMRProfile,
-)
+import os
+LAMBDA_DIR = os.path.abspath(os.path.join(os.path.dirname(os.path.abspath(__file__)), '../../lambda_sources/'))
 
-from .cluster_configuration import (
-    BaseConfiguration,
-    InstanceGroupConfiguration
-)
+
+def _lambda_path(path):
+    return os.path.join(LAMBDA_DIR, path)