+++ aws_emr_launch/constructs/step_functions/emr_launch_function.py
@@ -26,7 +26,9 @@ from aws_cdk import (
 )
 
 from .emr_fragments import EMRFragments
-from ..emr_constructs.cluster_configurations import BaseConfiguration
+from ..emr_constructs.cluster_configuration import BaseConfiguration
+
+SSM_PARAMETER_PREFIX = '/emr_launch/emr_launch_functions'
 
 
 class EMRLaunchFunctionNotFoundError(Exception):
@@ -37,6 +39,7 @@ class EMRLaunchFunction(core.Construct):
     def __init__(self, scope: core.Construct, id: str, *,
                  cluster_config: Optional[BaseConfiguration] = None,
                  launch_function_name: Optional[str] = None,
+                 namespace: str = 'default',
                  default_fail_if_job_running: bool = False,
                  success_topic: Optional[sns.Topic] = None,
                  failure_topic: Optional[sns.Topic] = None,
@@ -57,7 +60,7 @@ class EMRLaunchFunction(core.Construct):
         fail_if_job_running_task = EMRFragments.fail_if_job_running_task(
             self, default_fail_if_job_running=default_fail_if_job_running)
 
-        run_job_flow_task = EMRFragments.run_job_flow_task(self)
+        create_cluster_task = EMRFragments.create_cluster_task(self)
 
         fail = EMRFragments.fail_fragment(
             self,
@@ -74,7 +77,7 @@ class EMRLaunchFunction(core.Construct):
         definition = \
             override_cluster_configs_task.add_catch(fail, errors=['States.ALL'], result_path='$.Error') \
             .next(fail_if_job_running_task.add_catch(fail, errors=['States.ALL'], result_path='$.Error')) \
-            .next(run_job_flow_task.add_catch(fail, errors=['States.ALL'], result_path='$.Error')) \
+            .next(create_cluster_task.add_catch(fail, errors=['States.ALL'], result_path='$.Error')) \
             .next(success)
 
         self._state_machine = sfn.StateMachine(
@@ -85,11 +88,19 @@ class EMRLaunchFunction(core.Construct):
             self._ssm_parameter = ssm.StringParameter(
                 self, 'SSMParameter',
                 string_value=json.dumps({
+                    'LaunchFunctionName': launch_function_name,
+                    'ClusterConfiguration': cluster_config.cluster_name,
+                    'DefaultFailIfJobRunning': default_fail_if_job_running,
+                    'SuccessTopic': success_topic.topic_arn if success_topic is not None else None,
+                    'FailureTopic': failure_topic.topic_arn if failure_topic is not None else None,
+                    'OverrideClusterConfigsLambda':
+                        override_cluster_configs_lambda.function_arn
+                        if override_cluster_configs_lambda is not None
+                        else None,
                     'AllowedClusterConfigOverrides': self._allowed_cluster_config_overrides,
-                    'FunctionArn': self._state_machine.state_machine_arn
+                    'StateMachineArn': self._state_machine.state_machine_arn
                 }),
-                parameter_name='/emr_launch/control_plane/emr_launch_functions/{}'.format(
-                    launch_function_name))
+                parameter_name=f'{SSM_PARAMETER_PREFIX}/{namespace}/{launch_function_name}')
 
     @property
     def allowed_cluster_config_overrides(self) -> Mapping[str, str]:
@@ -100,17 +111,27 @@ class EMRLaunchFunction(core.Construct):
         return self._state_machine
 
     @staticmethod
-    def from_stored_config(scope: core.Construct, id: str, launch_function_name: str):
+    def get_functions(namespace: str = 'default', next_token: Optional[str] = None) -> Mapping[str, any]:
+        params = {
+            'Path': f'{SSM_PARAMETER_PREFIX}/{namespace}/'
+        }
+        if next_token:
+            params['NextToken'] = next_token
+        result = boto3.client('ssm').get_parameters_by_path(**params)
+
+        functions = {
+            'EMRLaunchFunctions': [json.loads(p['Value']) for p in result['Parameters']]
+        }
+        if 'NextToken' in result:
+            functions['NextToken'] = result['NextToken']
+        return functions
+
+    @staticmethod
+    def get_function(launch_function_name: str, namespace: str = 'default') -> Mapping[str, any]:
         try:
-            function_json = boto3.client('ssm', region_name=core.Stack.of(scope).region).get_parameter(
-                Name='/emr_launch/control_plane/emr_launch_functions/{}'.format(
-                    launch_function_name))['Parameter']['Value']
-            launch_function = EMRLaunchFunction(scope, id)
-            stored_config = json.loads(function_json)
-            launch_function._allowed_cluster_config_overrides = stored_config['AllowedClusterConfigOverrides']
-            launch_function._state_machine = sfn.StateMachine.from_state_machine_arn(
-                launch_function, 'StateMachine', stored_config['FunctionArn'])
-            return launch_function
+            function_json = boto3.client('ssm').get_parameter(
+                Name=f'{SSM_PARAMETER_PREFIX}/{namespace}/{launch_function_name}')['Parameter']['Value']
+            return json.loads(function_json)
         except ClientError as e:
             if e.response['Error']['Code'] == 'ParameterNotFound':
                 raise EMRLaunchFunctionNotFoundError()