+++ aws_emr_launch/lambda_sources/__init__.py
@@ -11,11 +11,3 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-from .emr_profile import (
-    EMRProfile,
-)
-
-from .cluster_configuration import (
-    BaseConfiguration,
-    InstanceGroupConfiguration
-)