+++ aws_emr_launch/lambda_sources/emr_utilities/override_cluster_configs.py
@@ -18,8 +18,6 @@ import traceback
 
 from dictor import dictor
 
-from utils import *
-
 emr = boto3.client('emr')
 
 LOGGER = logging.getLogger()